protected-qr/python-core/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
numpy==2.0.2
opencv-python-headless==4.10.0.84
pillow==10.4.0
qrcode==7.4.2